Update vm_workload_consolidation job
the reason which changes archive policy "bool" to "low" is that
the "bool" rule does not contain the default "mean" method.

At the same time, the "low" archive policy does not contain "last"
aggregation method, so we modify the default value of the
aggregation method to "mean" which supports all the strategies.

The parameter metrics have been added to make it easier
to modify the metrics for each strategy.

reference docs:https://gnocchi.xyz/stable_3.0/rest.html#archive-policy

Change-Id: I3c356b36911b547c0fb419ac6b0a1294a14cc853