Check node trait before creating server
Placement may fail to update trait because of Conflict
the trait may be updated by the Nova compute
update_available_resource periodic task.
We need node status is enabled, so we check the node
trait and delay if it is not the correct status.
the max delay time is 10 minutes.

Depends-on: Ie98e2c000568ecac63e8981a8b7087029c0d3705

Change-Id: I3e45d4a66a6e1bf55499def8550da38ddf01b638