skip metric injection when not needed
This patch is adding a new option `inject` into the method
_create_one_instance_per_host_with_statistic to specify if metrics
should be injected or not for the new instances.

The patch also introduces the new option in all the calls to the method
from the different tests. Given that different tests in the same testing
class may have different requirements, I'm doing it per-test and using a
constant in the test for clarity.

This patch is also removing a skip condition in the real-data tests based
on the optimize.datasource configuration which is not longer needed.

Finally, I am adding job watcher-prometheus-integration to the check
pipeline so that changes in the plugin are properly gated with
prometheus backend.

Depends-On: https://review.opendev.org/c/openstack/watcher/+/945627

Change-Id: I6e02062ea01240702f7f09db5a22005c43e92ee5