Add support for prometheus datasource in scenario tests
Adds a new datasource option to be selected: prometheus
Adds methods to generate dake metrics and push them to
a prometheus server. Metrics supported are: host cpu metrics
and instance cpu metrics.
Adds a prometheus client based on subprocess call which
depends on promtool installed in the same host that tempest
is running. Additional support to choose a proxy host can
be added as a follow-up.

Change-Id: I407bb41560716df2ae73a2c44b908ae4ca27f254