Add support to run promtool in a podified deployment
Promtool client now supports running commands in a remote
host and in a podified deployment.
To run promtool in a podified environment, the host (or
a proxy host) needs a 'oc' installed and configured to communicate
with the OpenShift cluster, in order to reach Prometheus service
pods. New configs options where added to allow access to a proxy
host and to communicate with the OpenShift cluster. A config option
was added to define the type of OpenStack deployment: devstack or
podified

Change-Id: I74353318cded468a139a94cc9cae79615cb04f6e