Check expected actions included in plans when executing strategies
Currently, if a strategy execution creates an empty plan, the test
passes. While that may be a valid result in some cases, in other it may
hide problems, and the plugin should check that the expected actions
are created.

This patch introduces a new parameter expected_actions in the
execute_actions method to check that the all the action types expected
are found in the resulting plan. By default, this is set to None
and no check is done.

I'm also introducing the check in the test_execute_dummy_strategy for
reference, although, we will introduce it in more relevant jobs as
follow up patches (checking at the logs, some of the execute tests may
need also to be patches to actually generate non-empty plans).

Change-Id: I2c6a865bf682ff58b8f06a307a27eb599b365f7c