Add real-data scenario tests
Currently, we are injecting metrics to run some scenarios. While, that
provides flexibility and it is a good way for many cases, we still want
to have some jobs which create real workload to test the end to end
workflow. A new parameter is added real_workload_period is created to define
the duration of the created workload.

This patch is adding a new test class TestRealExecuteStrategies for
strategies testing with real data.

Note that these test are mutually exclusive, as faked metrics may lead
to unexpected behavior.

The first test in the class is using workload_stabilization strategy
based on instance_cpu_usage workload.

In order to make easier to disable running these tests, I'm adding an
annotation with types slow and real_load. I will filter out those tests
by default in the jobs definition in watcher repo.

Change-Id: I910878bacc6e7964ef189d471bf6b612bfaacf26