Enforce compute host in create_one_instance_per_host
Currently, it relies in the nova scheduler to schecule each instance in
different hosts. While this leads to the expected behaviro in most cases
I have seen cases where both instances are assigned to the same node.

To avoid this issue, this patch is enfocing the host name assignment
when creating the instances.

Note that this patch requires nova api version to be at least 2.74 [1].
I am setting the parameter compute_min_microversion at the test class
level which defines the minimal version of nova required and used
when calling Nova API. For a cluster which does not provides it, it will
skip the tests requiring it, which are those using
_create_one_instance_per_host_with_statistic function.

Given that 2.74 version is from Train [2], I am not providing an alternative
implementation for those cases, other than skipping the tests.

Note that this patch does not affect to the version of the nova api that
watcher itself uses which is defined by a parameter in watcher config [3].

Finally, this patch is also fixing the call to skip_checks of the parent
class which implements skipping tests based on the nova api version.

[1] https://docs.openstack.org/api-ref/compute/#servers-servers
[2] https://docs.openstack.org/nova/latest/reference/api-microversion-history.html#id67
[3] https://docs.openstack.org/watcher/latest/configuration/watcher.html#nova_client.api_version

Change-Id: Icc28aed488a992920601fab675b9d30aacac6e96