Change timeframe of fake metrics uploaded to gnocchi.
I have detected that the test errors on this strategy are related to inyecting fake metrics on gnocchi.

When we create the fake metrics from now until 5 minutes in the past, then we sometimes move the instances to other hypervisor and wait for the model. If this process takes one or two minutes, when decision engine collects the metrics, gnocchi execute a rate between the start point and the end. if metric is empty at the end, collect fails and decision engine set the load to 0

Moving the metrics 2 minutes to the future allow us to always get valid metrics

Also updating the threshold parameters on test_execute_workload_stabilization_strategy to match the real data one

Change-Id: I7493ca2ed9e8188b7537b7b2f3353f10266e972c
Closes-Bug: #2110107