Move function for creating instances from 1 to variable number
Current behavior for creating multiple instances per host iterating over
create_one_instance_per_host function is working fine because this block
https://opendev.org/openstack/watcher-tempest-plugin/src/commit/8aa31042302740929043f866e5540cf95af3ebf2/watcher_tempest_plugin/tests/scenario/base.py#L263
never return anything, as it is executed as non admin and instances are
on a different project.

If any time we change the client to be admin, like on
https://review.opendev.org/c/openstack/watcher-tempest-plugin/+/950389,
iterating over create_one_instance_per_host will fail.

That why Im moving this function to a create a variable number of
instances per host

Change-Id: Iacc8f05fb3590151ab1609330c719729315c7fec