Fix injected host_ram_usage metrics
We are currently injecting them in MB, but the acual metrics on the
gnocchi and prometheus backends store those metrics in other units:

- For gnocchi, it's using ceilometer host metrics and
  hardware.memory.used [1].
- For prometheus, it is calculated from node_memory_MemAvailable_bytes
  which is in bytes.

This patch is adapting the injected metrics to those units.

I'm also cleaning the labels for the host ram metrics to the prometheus
backend which included cpu related labels.

[1] https://docs.openstack.org/ceilometer/train/admin/telemetry-measurements.html#snmp-based-meters

Change-Id: Iac905f48f27265a3c7152c3c6508b919478bc93d