Add custom flavor and dynamic threshold to workload_balance tests
As we can have different flavors on the different CIs validating
this code, we cannot rely on them to evaluate the strategies.

With this change, I create a flavor using 15% of the hypervisor RAM
where instances are going to be created, so we can easily assign
the threshold for memory utilization on workload balance strategy

Flavor is created and deleted during the execution.

For CPU tests, modifying the flavor is not a valid solution as usage is
calculated based on the number of vcpus of the hypervisor, so the
solution is to assign the threshold of the audit based on the number of
vcpus of the hypervisor

Change-Id: I2d4398fe5822e0bf35b95fc1a77d1e2553327e8c
Closes-Bug: https://bugs.launchpad.net/watcher-tempest-plugin/+bug/2115733
Signed-off-by: morenod <dsanzmor@redhat.com>