Add support for microversion testing for api and scenario tests
Following the tempest guide[1] for implementing microversion tests,
a couple of changes were made in the plugin to better adapt it to support
new test classes for different microversions. When not defined a minimum
microversion, the min microversion from the service will be used.
This change only affects the api and scenarios tests, since the
client_functional tests have their own way to setup the api version,
which can be further improved if needed.
Some test classes had to set a minimum api version due to helper functions
like the use of `GET /v1/data_model`.
Two new configuration options were added to define the test target
microversions, and should be used to help on testing stable branches
in future releases.

[1] https://docs.openstack.org/tempest/latest/microversion_testing.html#how-to-implement-microversion-tests

Depends-On: https://review.opendev.org/c/openstack/watcher/+/956380

Change-Id: Id8fa25593b5cd47f5943b228415cd191e41e19ab
Signed-off-by: Douglas Viroel <viroel@gmail.com>