Remove server_state check from data_model test
Remove the validation of server_state in the data_model test,
since the test can't guarantee that a notification was already
processed and the state updated from building to active. This
makes the job to fail in some runs. We don't actually need to validate
this state, which not always represents the real state of the
instance.

Change-Id: I3f87db1af19ffbef4b3fad23a3b01ea122eeccde
Signed-off-by: Douglas Viroel <viroel@gmail.com>