Refresh the state of the actionplan before asserting status
This is follow-up of [1]. That patch added a wait for actionplan to
reach finished state before checking the status. However, it is not
refreshing the state of the actionplan in the action_plan variable
before doint the actual assert, so it's still validating the state that
had before the wait.

[1] https://review.opendev.org/c/openstack/watcher-tempest-plugin/+/968750

Closes-Bug: #2133181

Change-Id: Id3520d9293778694ee699be4b0b4ab07144dae3a
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>